# storage/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a server for the storage of expressions and their results",
        "title": "Swagger Storage API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1.",
    "paths": {
        "/confirmStartCalculating": {
            "post": {
                "description": "Confirm start calculating for expression to coordinate work of calculation servers",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "updates (used by calculation server)"
                ],
                "summary": "Confirm start calculating",
                "parameters": [
                    {
                        "description": "Expression",
                        "name": "expression",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.InConfirmStartOfCalculating"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.OutConfirmStartOfCalculating"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.OutConfirmStartOfCalculating"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.OutConfirmStartOfCalculating"
                        }
                    }
                }
            }
        },
        "/expression": {
            "get": {
                "description": "Get all expressions from storage",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "expression"
                ],
                "summary": "Get all expressions",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.OutGetAllExpressions"
                        }
                    }
                }
            },
            "post": {
                "description": "Add expression to storage",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "expression"
                ],
                "summary": "Add expression",
                "parameters": [
                    {
                        "description": "Expression",
                        "name": "expression",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.InPostExpression"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.OutPostExpression"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.OutPostExpression"
                        }
                    }
                }
            }
        },
        "/expressionById": {
            "get": {
                "description": "Get expression from storage by id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "expression"
                ],
                "summary": "Get expression by id",
                "parameters": [
                    {
                        "description": "Expression ID",
                        "name": "id",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.InGetExpressionByID"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.OutGetExpressionByID"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.OutGetExpressionByID"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.OutGetExpressionByID"
                        }
                    }
                }
            }
        },
        "/getComputingPowers": {
            "get": {
                "description": "Get computing powers from storage",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "operations"
                ],
                "summary": "Get computing powers",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.OutGetComputingPowers"
                        }
                    }
                }
            }
        },
        "/getExpressionByServer": {
            "get": {
                "description": "Get expressions from storage by server name",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "expression"
                ],
                "summary": "Get expression by server",
                "parameters": [
                    {
                        "description": "Server name",
                        "name": "server_name",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.InGetExpressionByServer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.OutGetExpressionByServer"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.OutGetExpressionByServer"
                        }
                    }
                }
            }
        },
        "/getOperationsAndTimes": {
            "get": {
                "description": "Get operations and times for calculation as a map of operation and time in milliseconds, {\"+\": 100,...}",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "operations"
                ],
                "summary": "Get operations and times",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.OutGetOperationsAndTimes"
                        }
                    }
                }
            }
        },
        "/getUpdates": {
            "get": {
                "description": "Get not working expressions for calculation server",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "updates (used by calculation server)"
                ],
                "summary": "Get updates",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.OutGetUpdates"
                        }
                    }
                }
            }
        },
        "/keepAlive": {
            "post": {
                "description": "Keep alive for expression to coordinate work of calculation servers",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "updates (used by calculation server)"
                ],
                "summary": "Keep alive",
                "parameters": [
                    {
                        "description": "Expression",
                        "name": "expression",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.InKeepAlive"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.OutPing"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.OutPing"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.OutPing"
                        }
                    }
                }
            }
        },
        "/ping": {
            "get": {
                "description": "Check connection with server",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ping"
                ],
                "summary": "Ping",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.OutPing"
                        }
                    }
                }
            }
        },
        "/postOperationsAndTimes": {
            "post": {
                "description": "Set operations and times for calculation as a map of operation and time in milliseconds, {\"+\": 100,...}",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "operations"
                ],
                "summary": "Set operations and times",
                "parameters": [
                    {
                        "description": "Operations and times",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "integer"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.OutSetOperationsAndTimes"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.OutSetOperationsAndTimes"
                        }
                    }
                }
            }
        },
        "/postResult": {
            "post": {
                "description": "Post result of the calculation",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "updates (used by calculation server)"
                ],
                "summary": "Post result",
                "parameters": [
                    {
                        "description": "Expression",
                        "name": "expression",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.InPostResult"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.OutPostResult"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.OutPostResult"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.OutPostResult"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "api.InConfirmStartOfCalculating": {
            "type": "object",
            "required": [
                "expression"
            ],
            "properties": {
                "expression": {
                    "$ref": "#/definitions/db.Expression"
                }
            }
        },
        "api.InGetExpressionByID": {
            "type": "object",
            "required": [
                "id"
            ],
            "properties": {
                "id": {
                    "type": "integer"
                }
            }
        },
        "api.InGetExpressionByServer": {
            "type": "object",
            "required": [
                "server_name"
            ],
            "properties": {
                "server_name": {
                    "type": "string"
                }
            }
        },
        "api.InKeepAlive": {
            "type": "object",
            "required": [
                "expression"
            ],
            "properties": {
                "expression": {
                    "$ref": "#/definitions/db.Expression"
                }
            }
        },
        "api.InPostExpression": {
            "type": "object",
            "required": [
                "expression"
            ],
            "properties": {
                "expression": {
                    "type": "string"
                }
            }
        },
        "api.InPostResult": {
            "type": "object",
            "required": [
                "expression"
            ],
            "properties": {
                "expression": {
                    "$ref": "#/definitions/db.Expression"
                }
            }
        },
        "api.OutConfirmStartOfCalculating": {
            "type": "object",
            "properties": {
                "confirm": {
                    "type": "boolean"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "api.OutGetAllExpressions": {
            "type": "object",
            "properties": {
                "expressions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/db.Expression"
                    }
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "api.OutGetComputingPowers": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "type": "object",
                        "properties": {
                            "calculated_expressions": {
                                "type": "array",
                                "items": {
                                    "type": "integer"
                                }
                            },
                            "server_name": {
                                "type": "string"
                            }
                        }
                    }
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "api.OutGetExpressionByID": {
            "type": "object",
            "properties": {
                "expression": {
                    "$ref": "#/definitions/db.Expression"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "api.OutGetExpressionByServer": {
            "type": "object",
            "properties": {
                "expressions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/db.Expression"
                    }
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "api.OutGetOperationsAndTimes": {
            "type": "object",
            "properties": {
                "data": {
                    "description": "executions times in milliseconds: {\"+\": 100,...}",
                    "type": "object",
                    "additionalProperties": {
                        "type": "integer"
                    }
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "api.OutGetUpdates": {
            "type": "object",
            "required": [
                "tasks"
            ],
            "properties": {
                "message": {
                    "type": "string"
                },
                "tasks": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/db.Expression"
                    }
                }
            }
        },
        "api.OutPing": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
        "api.OutPostExpression": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "api.OutPostResult": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
        "api.OutSetOperationsAndTimes": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
        "db.Expression": {
            "type": "object",
            "properties": {
                "alive_experise_at": {
                    "type": "integer"
                },
                "answer": {
                    "type": "number"
                },
                "creation_time": {
                    "type": "string"
                },
                "end_calculation_time": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "logs": {
                    "type": "string"
                },
                "ready": {
                    "description": "0 - not ready, 1 - working, 2 - ready, 3 - error",
                    "type": "integer"
                },
                "server_name": {
                    "type": "string"
                },
                "value": {
                    "type": "string"
                }
            }
        }
    }
}